// multisto/src/Schema.ts
export interface ApiMigrationTable {
  id: number;
  timestamp: number;
  name: string;
}
// ...
export interface Migrations {
  id: number;
  name: string;
  run_on: Date;
}
// ...
export interface Params {
  ID: number;
  param?: string | null;
  isglobal?: number | null;
  datatype?: number | null;
  stringValue?: string | null;
  intValue?: number | null;
}
// ...
export interface Stos {
  ID: number;
  title: string;
  details: string;
  /**  Defaults to: 1. */
  isActive?: number | null;
  logo: string;
  ethereumContractAddress: string;
  ethereumWhitelistAddress: string;
  disclamer?: string | null;
  stolink?: string | null;
  stolinkfull?: string | null;
  /**  Defaults to: 0. */
  stoType?: number | null;
  stoinvestortypes?: string | null;
  emailFooter?: string | null;
  steps?: string | null;
  registrationtext?: string | null;
  SMTP_Host?: string | null;
  SMTP_Port?: string | null;
  SMTP_User?: string | null;
  SMTP_Password?: string | null;
  SMTP_FromAddress?: string | null;
  website?: string | null;
  stoinvestortypesnotonshareregister?: string | null;
  /**  Defaults to: 0. */
  companytype: number;
  settings?: string | null;
  registrationsuccesstext?: string | null;
  tellafriendtext?: string | null;
  inviteFriendEmailText?: string | null;
  PropertyFullDetails?: string | null;
  /**  Defaults to: 2030-01-01. */
  exchangeOpenDate?: Date | null;
  propertypicture?: string | null;
  docusign_sto_contract?: string | null;
  docusign_sto_purchase?: string | null;
  /**  Defaults to: 0. */
  externalSystemID?: number | null;
  projectAddress?: string | null;
  LegalDetails?: string | null;
  affiliatePlanId?: number | null;
  affiliateShareTypeId?: number | null;
  /**  Defaults to: 0. */
  isICOShareTypeCompany?: number | null;
  EmailTxtInvestorBulkUpload?: string | null;
  /**  Defaults to: 1. */
  isBuyButtonEnabled?: number | null;
  isBimountEnabled?: number | null;
  /**  Defaults to: 0. */
  projectCost?: number | null;
  VerifyInvestorComHostToken?: string | null;
  helloSignClientID?: string | null;
  SMTP_FromName?: string | null;
  /**  Defaults to: CURRENT_TIMESTAMP. */
  createdAt: Date;
  /**  Defaults to: 0. */
  defaultBlockchain?: number | null;
}